Switch to `posix_spawn` for process creation (experimental)
Benchmarks:
 - For 1000 processes (where DEmOS uses ~30 MB RAM), fork+exec and posix_spawn are approximately equally fast.
 - By allocating a 1GB buffer in DEmOS and filling it with garbage, fork+exec slows down considerably. As expected, posix_spawn doesn't care.
 - For 10k processes, both die, because i.MX8 only has 6GB RAM
  (~0.6MB available for each process) and kernel invokes OOM killer somewhere during startup.

According to a few measurements I've done, for the expected workloads, I don't think there's a major difference - CoW mostly saves the day for fork+exec, and DEmOS doesn't typically take up enough memory for the page table setup overhead to become visible. The setup for fork+exec is easier to adjust for our needs, so it's probably better to keep it.

I'll push this to a new branch and we'll see.